Refactor common component
