+++ src/components/Visualizations/GraphPanes/AuthorNetworkGraphPane.tsx
@@ -1,7 +1,8 @@
 import { IADSApiAuthorNetworkNode, IADSApiAuthorNetworkNodeKey } from '@api';
-import { Radio, RadioGroup, Stack, Text, Input, Button, FormControl, FormLabel, Switch } from '@chakra-ui/react';
-import { ChangeEvent, KeyboardEvent, ReactElement, useState } from 'react';
+import { Radio, RadioGroup, Stack, FormControl, FormLabel, Switch } from '@chakra-ui/react';
+import { ChangeEvent, ReactElement, useState } from 'react';
 import { AuthorNetworkGraph } from '../Graphs/AuthorNetworkGraph';
+import { PaperLimit } from '../Panes';
 import { IView } from './types';
 
 export interface IAuthorNetworkGraphPaneProps {
@@ -42,7 +43,7 @@ export const AuthorNetworkGraphPane = ({
 
   return (
     <Stack as="section" aria-label="Author Network" width="100%" mt={5}>
-      <LimitPaper initialLimit={paperLimit} max={maxPaperLimit} onApply={handleChangePaperLimit} />
+      <PaperLimit initialLimit={paperLimit} max={maxPaperLimit} onApply={handleChangePaperLimit} />
       <RadioGroup defaultChecked onChange={handleChangeView} value={view.id}>
         <Stack direction="row">
           {views.map((v) => (
@@ -60,47 +61,6 @@ export const AuthorNetworkGraphPane = ({
         onClickNode={onClickNode}
         keyToUseAsValue={view.valueToUse as IADSApiAuthorNetworkNodeKey}
       />
-    </Stack>
-  );
-};
-
-const LimitPaper = ({
-  initialLimit,
-  max,
-  onApply,
-}: {
-  initialLimit: number;
-  max: number;
-  onApply: (n: number) => void;
-}) => {
-  const [limit, setLimit] = useState(initialLimit);
-
-  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
-    const value = parseInt(e.target.value);
-    setLimit(value);
-  };
-
-  const handleKeydown = (e: KeyboardEvent<HTMLInputElement>) => {
-    if (e.key === 'Enter') {
-      handleApply();
-    }
-  };
-
-  const handleApply = () => {
-    if (isNaN(limit) || limit < 1 || limit > max) {
-      setLimit(initialLimit);
-      onApply(initialLimit);
-    } else {
-      onApply(limit);
-    }
-  };
-
-  return (
-    <Stack direction="row" alignItems="center" my={2}>
-      <Text>Show the first</Text>
-      <Input w={16} type="number" value={limit} max={max} onChange={handleChange} onKeyDown={handleKeydown} />
-      <Text>{`papers (max is ${max})`}</Text>
-      <Button onClick={handleApply}>Apply</Button>
     </Stack>
   );
 };
