+++ src/components/Visualizations/Panes/PaperLimit.tsx
@@ -1,0 +1,43 @@
+import { Stack, Input, Button, Text } from '@chakra-ui/react';
+import { ChangeEvent, ReactElement, useState, KeyboardEvent } from 'react';
+
+export const PaperLimit = ({
+  initialLimit,
+  max,
+  onApply,
+}: {
+  initialLimit: number;
+  max: number;
+  onApply: (n: number) => void;
+}): ReactElement => {
+  const [limit, setLimit] = useState(initialLimit);
+
+  const handleChange = (e: ChangeEvent<HTMLInputElement>) => {
+    const value = parseInt(e.target.value);
+    setLimit(value);
+  };
+
+  const handleKeydown = (e: KeyboardEvent<HTMLInputElement>) => {
+    if (e.key === 'Enter') {
+      handleApply();
+    }
+  };
+
+  const handleApply = () => {
+    if (isNaN(limit) || limit < 1 || limit > max) {
+      setLimit(initialLimit);
+      onApply(initialLimit);
+    } else {
+      onApply(limit);
+    }
+  };
+
+  return (
+    <Stack direction="row" alignItems="center" my={2}>
+      <Text>Show the first</Text>
+      <Input w={16} type="number" value={limit} max={max} onChange={handleChange} onKeyDown={handleKeydown} />
+      <Text>{`papers (max is ${max})`}</Text>
+      <Button onClick={handleApply}>Apply</Button>
+    </Stack>
+  );
+};
