+++ src/components/Visualizations/Panes/index.ts
@@ -1,1 +1,2 @@
 export * from './NetworkDetailsPane';
+export * from './PaperLimit';
